+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Delete tag:   git tag -d <tag_name>
 
 ## [Unreleased]
 ### Fixed
+- Fixed console colors
+- Fixed console input to accept input without options
 - Fixed console shell options now show on sub commands
 - Fixed console showing help when using multiple required arguments
 
@@ -28,6 +30,7 @@ Delete tag:   git tag -d <tag_name>
 - Added custom meta mapping and changed model results to use this instead.
 
 ### Added
+- Added shell:notice shell:info shell:debug
 - ConfigTrait delete if null passed
 - DB console shell app
 - Plugin install shell command